Didn't change any code, made Warehouse pages into components and refactored my pages using the components
