+++ src/pages/Inventories/Inventories.jsx
